* develop: (68 commits)
  Update CHANGELOG [ci skip]
  Fix of wrong include symbol of CCVector.h
  closed #3290: [JSB] The order of onEnter and onExit is wrong. For develop branch.
  Update CHANGELOG [ci skip]
  Update AUTHORS [ci skip]
  Added test case for binary fnt file
  [AUTO] : updating submodule reference to latest autogenerated bindings
  Small fix for `FREE` wrong variable.
  More warning fixes.
  AngelCode binary file format support for LabelBMFont
  [AUTO] : updating submodule reference to latest autogenerated bindings
  Bug fixes for cocostudio.
  Update AUTHORS [ci skip]
  Update CHANGELOG [ci skip]
  Improve the coding style
  Modify wrong codes
  [AUTO] : updating submodule reference to latest autogenerated bindings
  fix failed to receive dispatch signal on some android device,such as [HTC EVO 3D X515m].
  Fixed bugs
  [AUTO] : updating submodule reference to latest autogenerated bindings
  ...

+++ cocos/2d/CCDrawingPrimitives.cpp
@@ -475,7 +475,7 @@ void drawCardinalSpline( PointArray *config, float tension,  unsigned int segmen
 
     Vertex2F* vertices = new Vertex2F[segments + 1];
 
-    unsigned int p;
+    ssize_t p;
     float lt;
     float deltaT = 1.0f / config->count();
