+++ cocos/2d/CCMenu.cpp
@@ -346,7 +346,7 @@ void Menu::alignItemsInColumns(int columns, va_list args)
 void Menu::alignItemsInColumnsWithArray(const ValueVector& rows)
 {
     int height = -5;
-    int row = 0;
+    size_t row = 0;
     int rowHeight = 0;
     int columnsOccupied = 0;
     int rowColumns = 0;
@@ -441,7 +441,7 @@ void Menu::alignItemsInRowsWithArray(const ValueVector& columns)
 
     int width = -10;
     int columnHeight = -5;
-    int column = 0;
+    size_t column = 0;
     int columnWidth = 0;
     int rowsOccupied = 0;
     int columnRows;
